adding InteropGroupPolicyReader interface
