+++ libs/membership/membership-common/build.gradle
@@ -6,6 +6,7 @@ plugins {
 description 'Membership Internal API'
 
 dependencies {
+    compileOnly 'org.osgi:org.osgi.service.component.annotations'
     compileOnly "org.osgi:osgi.annotation"
 
     implementation platform("net.corda:corda-api:$cordaApiVersion")
